requirements.txt
glfw==2.7.0
numpy==1.26.4
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyrr==0.10.3
pillow==10.4.0
pyyaml
scikit-learn==1.3.0
joblib==1.4.2
pyqt5
matplotlib
opencv-python
easing_functions
numba